refactor: update all controllers and services logic

+++ src/main/java/ru/se/ifmo/tinder/repository/FabricTextureRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.se.ifmo.tinder.model.FabricTexture;
 
 @Repository
-public interface FabricTextureRepository extends JpaRepository<FabricTexture, Integer> {
+public interface FabricTextureRepository extends JpaRepository<FabricTexture, Long> {
 }
