+++ src/main/java/ru/se/ifmo/tinder/service/exceptions/NoEntityWithSuchIdException.java
@@ -3,15 +3,15 @@ package ru.se.ifmo.tinder.service.exceptions;
 import java.util.List;
 
 public class NoEntityWithSuchIdException extends RuntimeException {
-    public NoEntityWithSuchIdException(String name, Integer id) {
+    public NoEntityWithSuchIdException(String name, Long id) {
         super("%s with id: %d not found".formatted(name, id));
     }
 
-    public NoEntityWithSuchIdException(String name, String field, Integer id) {
+    public NoEntityWithSuchIdException(String name, String field, Long id) {
         super("%s with %s id: %d not found".formatted(name, field, id));
     }
 
-    public NoEntityWithSuchIdException(String name, List<Integer> ids) {
+    public NoEntityWithSuchIdException(String name, List<Long> ids) {
         super("%s with ids: %s not found".formatted(name, ids));
     }
 }
